+++ float/src/fbig_cached_ops.rs
@@ -660,6 +660,7 @@ impl<R: ErrorBounds, const B: Word> CachedFBig<R, B> {
     forward_to_fbig!(powi(exp: dashu_int::IBig));
     forward_to_context!(ln);
     forward_to_context!(ln_1p);
+    forward_to_context!(log2);
     forward_to_context!(exp);
     forward_to_context!(exp_m1);